+++ src/modules/bookmark/dto/request/essayIdsReq.dto.ts
